modularize eerie sound and optionsP2
