+++ eerieSound.js
@@ -1,0 +1,9 @@
+const playEerieSound = () => {
+  $("#eerie")[0].play();
+};
+
+const muteEerieSound = () => {
+  $("#eerie").prop("muted", true);
+};
+
+export { playEerieSound, muteEerieSound };
